requirements.txt
streamlit>=1.27.0
python-docx
docx2pdf
Pillow==10.2.0
python-dateutil==2.8.2
